InputKeyboard class more generic and in own source files
It now has virtual OnKeyPress() and OnKeyRelease() methods that a
derived class can override to react to those event.

So in keyboardwrapper.cpp there now is a very short class derived
from InputKeyboard that passes key presses/releases on to the
EmulatedKeyboard in in those (overriden) methods.

In contrast to the EmulatedKeyboard class which I license under
MIT license (like NicoHood's code it's based on), InputKeyboard is
under GPLv2, like the USB Host Shield Library 2.0 which is needed to
use it. Probably doesn't make much sense to use a different license
as you'll have to respect the GPLv2 anyway when using this class.